move things to do with binding the service to their own package
